Create User class, fetch user from localStorage
I'd like to enable persistence for a user, for the purposes of saving
data during Demo.

The User class has two properties:
  name: string
  education: Education[]

It also has an init method, which attempts to retrieve the user
from localStorage. If the user doesn't exist, then it saves the new
user in localStorage.

There will be future methods to the User class added such as
saveEducation, editEducation, deleteEducation, etc.